+++ RayTracing/Material.java
@@ -4,15 +4,15 @@ public class Material {
 	final Color diffuse;
 	final Color specular;
 	final Color reflection;
-	final float phong;
-	final float trans;
+	final double phong;
+	final double trans;
 
 	Material(
 		double dr, double dg, double db,
 		double sr, double sg, double sb,
 		double rr, double rg, double rb,
-		float phong,
-		float trans) {
+		double phong,
+		double trans) {
 
 		this.diffuse = new Color(dr, dg, db);
 		this.specular = new Color(sr, sg, sb);
